Adds Logic for finding a placement for item
Adds a method to perform a line trace and if it hits it adds a another
instance of the carried item in that location. Will be used to place a
"ghost" version of the item to indicate to the player where the item
they place will end up. The intention is the angle the line trace is
shot out at is adjustable by the player to change height.