Adds first working version of item movement
Player can press 'E' key to pickup an item and will place it in a
location in front of them. There is also the groundwork for displaying a
ghost object, but that is not quite working.

Needed are some checks to make sure object isn't placed within geometry
and needs input to allow rotations and placing things above a certain
angle.