Trace for collision to prevent clipping.
When setting actor location in the ItemPlacementComponent we now trace
to see if there is a collision in an effort to prevent clipping. However
that does have the side effect that the actor's location is not always
updated. This commit includes some commented out work in progress to
counteract this side effect.