Simpler system for placing items.
Overlaps will are now allowed for the ghost item, but the player will
not be able to place an object that is overlapping. Major downside is
that this affects placing objects on walls. Would like to revisit if
there is time. Future plan is to give the ghost item a material that
makes it clear when it is overlapping.